Cleaned up and separated out the KavitaPlus classes to be in the DTO folder. While internal only, it makes navigation a lot easier for me.
