+++ API/Controllers/LicenseController.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using API.Constants;
 using API.Data;
-using API.DTOs.License;
+using API.DTOs.KavitaPlus.License;
 using API.Entities.Enums;
 using API.Extensions;
 using API.Services;
