+++ API/DTOs/KavitaPlus/ExternalMetadata/SeriesDetailPlusApiDto.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using API.DTOs.Scrobbling;
+using API.DTOs.SeriesDetail;
+
+namespace API.DTOs.KavitaPlus.ExternalMetadata;
+
+internal class SeriesDetailPlusApiDto
+{
+    public IEnumerable<MediaRecommendationDto> Recommendations { get; set; }
+    public IEnumerable<UserReviewDto> Reviews { get; set; }
+    public IEnumerable<RatingDto> Ratings { get; set; }
+    public int? AniListId { get; set; }
+    public long? MalId { get; set; }
+}
