+++ API/DTOs/KavitaPlus/License/EncryptLicenseDto.cs
@@ -1,4 +1,4 @@
-﻿namespace API.DTOs.License;
+﻿namespace API.DTOs.KavitaPlus.License;
 
 public class EncryptLicenseDto
 {
