+++ API/DTOs/KavitaPlus/License/LicenseInfoDto.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace API.DTOs.License;
+namespace API.DTOs.KavitaPlus.License;
 
 public class LicenseInfoDto
 {
